Added 'only_tags' and 'except_tags' named scopes to Taggings, and bumped gem version.
This was written for a very specific reason: I needed banned tags, so I needed a way to destroy a bunch of tags, regardless of owner. The tag_list wasn't doing it. Not the cleanest solution, but it lets me do this:

foo.tag_taggables.with_tags(foo.banned_tags).destroy_all